Tighten tolerance for rotation test.
From 1e-3 to 1e-10.

Also check that the integrator keeps good kinematics up
to O(1.0e6) time steps. This takes about 20s to run on my
machine (CPU only), but it is a very stringent test that
gives much confidence on the time integrator, the kinematics,
and the material model. Also, because it depends on these
three things being correct, it is very sensitive to errors
in any of these.