Improved error handling in MPMD executable
Prior to this commit, if an exception was thrown during creation
of the MPMD_App, the Plato Engine MPMD run would hang.  The
creation of the MPM_App was/is done in a try-catch block, and the
executable would exit from the catch block.  The other performers
would arrive at the next handleExceptions() checkin and hang.  To
prevent this, the catch block no longer exits but sets the App
pointer to nullptr and calls the PlatoInterface Catch() function
where the errors are reported to the console and registered in
the error handler and execution resumes.  The function
Plato::Interface::registerPerformer(...) now accepts a nullptr
which causes another exception to be registered.  At the next
handleExceptions() checkin the MPMD run exits gracefully.

story: TO-898

boy scout rule: created safeExit() function in LGR_MPMD to reduce
code duplication.