Add basic skeleton for OTM meshless method.
This is just a placeholder for now, but it's better to try to
give it some structure from the beginning.

There is a CMake option LGR_ENABLE_OTM to enable the OTM sources.
If this option is OFF, the OTM sources are disabled by code guards.

This is an attempt to avoid polluting the rest of the code with
OTM specific functionality. The OTM sources are located within the
lgrtk/v3/otm directory.