OTM: Add penalty contact & initial support
- Add option to use penalty contact. This was just
to check that the contact formulation wasn't the source
of the suspicious rebound in the Taylor bar. Using
both displacement contact or penalty contact the
rebound remains.

- Add option to use an initial support size different
than the number of nodes per element from the initial
mesh. This also seems to have no effect on the
rebound from the Taylor bar.