Mitigate convergence problems in OTM shape function recalculations
Fix the beta parameter once and for all at the beginning of an
analysis. Also set it the same everywhere in space. This avoids
the complication of propagating gradient of beta in computation
of shape function derivatives.

It also helps an observed problem in the recomputation of the
shape functions, where they would get ill-conditioned due to the
beta parameter growing too large.